s/requirements.txt
# requirements.txt
Flask
Flask-Cors
requests
python-dotenv
blockfrost-python
pycardano